Improved restarting a stalled app
Restarting the app every day was a blunt tool for fixing potentially
unresponsive apps. This commit adds a check for a file that is updated
every time the photo app checks in with the host. If there is more than
an hour without the photo app checking in, the app is restarted.

This also adds the ability to pause the photoctl script. When debugging,
it's nice to be able to turn off the cron job without having to mess
with crontab. By adding a third state for photoctl, "PAUSE", the
automtic stop/start of the app can be turned off as needed with
togglepause.py.